move create_samplesheet_csv to rnafusion
